refactor(cli): extract terminal keyboard code to hermes_cli/terminal_keyboard.py
Move ~320 lines of terminal keyboard protocol handling out of cli.py into
a dedicated module. Inline trivial _delete_previous_word wrapper, flatten
_register_word_delete_keybindings with stacked decorators. Sets the base
for future refactoring to reduce cli.py LOC and improve maintainability.
