+++ tests/test_cli_keybindings.py
@@ -1,4 +1,4 @@
-"""Tests for Hermes CLI prompt-toolkit keybindings."""
+"""Tests for Hermes CLI terminal keyboard handling."""
 
 from __future__ import annotations
 
@@ -11,27 +11,28 @@ from prompt_toolkit.clipboard import InMemoryClipboard
 from prompt_toolkit.document import Document
 from prompt_toolkit.input.vt100_parser import Vt100Parser
 from prompt_toolkit.key_binding import KeyBindings
+from prompt_toolkit.keys import Keys
 
-import cli
+from hermes_cli import terminal_keyboard as tkbd
 
 
 def test_register_word_delete_keybindings_wires_terminal_sequences():
     kb = KeyBindings()
 
-    cli._register_word_delete_keybindings(kb)
+    tkbd.register_word_delete_keybindings(kb)
 
     assert len(kb.bindings) == 3
 
-    with patch.object(cli, "_delete_previous_word") as delete_previous_word:
+    with patch.object(tkbd, "BACKWARD_KILL_WORD") as backward_kill_word:
         event = object()
         for binding in kb.bindings:
             binding.handler(event)
 
-    assert delete_previous_word.call_count == 3
+    assert backward_kill_word.call_count == 3
 
 
 def test_parse_terminal_keyboard_capabilities_detects_kitty_and_modify_other_keys():
-    capabilities = cli._parse_terminal_keyboard_capabilities(
+    capabilities = tkbd.parse_capabilities(
         "\x1b[?1u\x1b[>4;2m\x1b[?64;1;2;6;9;15;18;21;22c"
     )
 
@@ -40,7 +41,7 @@ def test_parse_terminal_keyboard_capabilities_detects_kitty_and_modify_other_key
 
 
 def test_parse_terminal_keyboard_detection_result_preserves_non_probe_input():
-    result = cli._parse_terminal_keyboard_detection_result(
+    result = tkbd.parse_detection_result(
         "\x1b[?1uhello\x1b[>4;2m\x1b[?64;1;2;6;9c"
     )
 
@@ -50,30 +51,30 @@ def test_parse_terminal_keyboard_detection_result_preserves_non_probe_input():
 
 
 def test_select_terminal_keyboard_mode_prefers_kitty():
-    capabilities = cli._TerminalKeyboardCapabilities(
+    capabilities = tkbd.TerminalKeyboardCapabilities(
         kitty_supported=True,
         modify_other_keys_supported=True,
     )
 
-    assert cli._select_terminal_keyboard_mode(capabilities) == "kitty"
+    assert tkbd.select_mode(capabilities) == "kitty"
 
 
 def test_select_terminal_keyboard_mode_falls_back_to_modify_other_keys():
-    capabilities = cli._TerminalKeyboardCapabilities(
+    capabilities = tkbd.TerminalKeyboardCapabilities(
         kitty_supported=False,
         modify_other_keys_supported=True,
     )
 
-    assert cli._select_terminal_keyboard_mode(capabilities) == "modify_other_keys"
+    assert tkbd.select_mode(capabilities) == "modify_other_keys"
 
 
 @pytest.mark.parametrize(
     ("mode", "enable", "expected_sequence"),
     [
-        ("kitty", True, cli._KITTY_KEYBOARD_ENABLE),
-        ("kitty", False, cli._KITTY_KEYBOARD_DISABLE),
-        ("modify_other_keys", True, cli._MODIFY_OTHER_KEYS_ENABLE),
-        ("modify_other_keys", False, cli._MODIFY_OTHER_KEYS_DISABLE),
+        ("kitty", True, tkbd.KITTY_KEYBOARD_ENABLE),
+        ("kitty", False, tkbd.KITTY_KEYBOARD_DISABLE),
+        ("modify_other_keys", True, tkbd.MODIFY_OTHER_KEYS_ENABLE),
+        ("modify_other_keys", False, tkbd.MODIFY_OTHER_KEYS_DISABLE),
     ],
 )
 def test_set_terminal_keyboard_mode_emits_expected_sequence(
@@ -83,7 +84,7 @@ def test_set_terminal_keyboard_mode_emits_expected_sequence(
 ):
     written = []
 
-    cli._set_terminal_keyboard_mode(
+    tkbd.set_mode(
         mode,
         enable=enable,
         writer=written.append,
@@ -95,13 +96,13 @@ def test_set_terminal_keyboard_mode_emits_expected_sequence(
 def test_install_ctrl_backspace_sequences_registers_modified_keycodes():
     sequences = {}
 
-    cli._install_ctrl_backspace_input_sequences(
+    tkbd.install_ctrl_backspace_sequences(
         sequences,
         terminfo_backspace=None,
     )
 
-    for sequence in cli._CTRL_BACKSPACE_ESCAPE_SEQUENCES:
-        assert sequences[sequence] == cli._CTRL_BACKSPACE_KEYS
+    for sequence in tkbd.CTRL_BACKSPACE_ESCAPE_SEQUENCES:
+        assert sequences[sequence] == tkbd.CTRL_BACKSPACE_KEYS
 
 
 @pytest.mark.parametrize(
@@ -117,12 +118,12 @@ def test_install_ctrl_backspace_sequences_uses_terminfo_backspace_swap(
 ):
     sequences = {}
 
-    cli._install_ctrl_backspace_input_sequences(
+    tkbd.install_ctrl_backspace_sequences(
         sequences,
         terminfo_backspace=terminfo_backspace,
     )
 
-    assert sequences[ctrl_backspace_byte] == cli._CTRL_BACKSPACE_KEYS
+    assert sequences[ctrl_backspace_byte] == tkbd.CTRL_BACKSPACE_KEYS
 
 
 def test_vt100_parser_maps_ctrl_backspace_csi_u_sequence():
@@ -132,60 +133,60 @@ def test_vt100_parser_maps_ctrl_backspace_csi_u_sequence():
     parser.feed("\x1b[127;5u")
     parser.flush()
 
-    assert [press.key for press in key_presses] == list(cli._CTRL_BACKSPACE_KEYS)
+    assert [press.key for press in key_presses] == list(tkbd.CTRL_BACKSPACE_KEYS)
 
 
 def test_vt100_parser_maps_ctrl_b_kitty_sequence():
     key_presses = []
     parser = Vt100Parser(key_presses.append)
 
-    parser.feed(cli._kitty_key_sequence(ord("b"), 5))
+    parser.feed(tkbd.kitty_key_sequence(ord("b"), 5))
     parser.flush()
 
-    assert [press.key for press in key_presses] == [cli.Keys.ControlB]
+    assert [press.key for press in key_presses] == [Keys.ControlB]
 
 
 def test_vt100_parser_maps_alt_v_kitty_sequence():
     key_presses = []
     parser = Vt100Parser(key_presses.append)
 
-    parser.feed(cli._kitty_key_sequence(ord("v"), 3))
+    parser.feed(tkbd.kitty_key_sequence(ord("v"), 3))
     parser.flush()
 
-    assert [press.key for press in key_presses] == [cli.Keys.Escape, "v"]
+    assert [press.key for press in key_presses] == [Keys.Escape, "v"]
 
 
 def test_vt100_parser_maps_alt_enter_modify_other_keys_sequence():
     key_presses = []
     parser = Vt100Parser(key_presses.append)
 
-    parser.feed(cli._modify_other_keys_sequence(13, 3))
+    parser.feed(tkbd.modify_other_keys_sequence(13, 3))
     parser.flush()
 
-    assert [press.key for press in key_presses] == [cli.Keys.Escape, cli.Keys.ControlM]
+    assert [press.key for press in key_presses] == [Keys.Escape, Keys.ControlM]
 
 
 def test_vt100_parser_maps_ctrl_enter_kitty_sequence_to_control_j():
     key_presses = []
     parser = Vt100Parser(key_presses.append)
 
-    parser.feed(cli._kitty_key_sequence(13, 5))
+    parser.feed(tkbd.kitty_key_sequence(13, 5))
     parser.flush()
 
-    assert [press.key for press in key_presses] == [cli.Keys.ControlJ]
+    assert [press.key for press in key_presses] == [Keys.ControlJ]
 
 
 def test_vt100_parser_maps_ctrl_enter_modify_other_keys_sequence_to_control_j():
     key_presses = []
     parser = Vt100Parser(key_presses.append)
 
-    parser.feed(cli._modify_other_keys_sequence(13, 5))
+    parser.feed(tkbd.modify_other_keys_sequence(13, 5))
     parser.flush()
 
-    assert [press.key for press in key_presses] == [cli.Keys.ControlJ]
+    assert [press.key for press in key_presses] == [Keys.ControlJ]
 
 
-def test_delete_previous_word_uses_prompt_toolkit_word_boundaries():
+def test_backward_kill_word_uses_prompt_toolkit_word_boundaries():
     buffer = Buffer()
     buffer.set_document(Document("alpha beta", cursor_position=len("alpha beta")))
     clipboard = InMemoryClipboard()
@@ -199,7 +200,7 @@ def test_delete_previous_word_uses_prompt_toolkit_word_boundaries():
         ),
     )
 
-    cli._delete_previous_word(event)
+    tkbd.BACKWARD_KILL_WORD(event)
 
     assert buffer.text == "alpha "
     assert buffer.cursor_position == len("alpha ")
